// src/models/Employee.model.ts
/*Model pojo class for Employee data mapping */

export interface Employee {
    
    employeeId : number | null | undefined;
	firstname : string | null | undefined;
	lastname: string | null | undefined;
	designation: string | null | undefined;
	city: string | null | undefined;
	country: string | null | undefined;
	age: number | null | undefined;
}